Enables posting the sale data back to the API
Qunatity and ProductId comes frome the Frontend, wheras RetailPrice is going to be fetched from the database - subtotal tax and total is re-calculated in the API.

+++ RMDesktopUI.Library/Helpers/ConfigHelper.cs
@@ -5,12 +5,11 @@ namespace RMDesktopUI.Library.Helpers
     public class ConfigHelper : IConfigHelper
     {
         // Reading configuration information from the App.Config
-        public double GetTaxRate() 
+        public decimal GetTaxRate() 
         {
-
             var taxRateText = ConfigurationManager.AppSettings["taxRate"];
 
-            var isValidTaxRate = double.TryParse(taxRateText, out var output);
+            var isValidTaxRate = decimal.TryParse(taxRateText, out var output);
 
             if (isValidTaxRate == false)
             {
